requirements.txt
functions-framework==3.*
wreq
google-cloud-bigquery